# [TECH] Amélioration des performances du database builder