--- a/api/db/database-builder/factory/build-client-application.js
+++ b/api/db/database-builder/factory/build-client-application.js
@@ -1,6 +1,10 @@
 import { cryptoService } from '../../../src/shared/domain/services/crypto-service.js';
 import { databaseBuffer } from '../database-buffer.js';
 
+const DEFAULT_CLIENT_SECRET = 'super-secret';
+// eslint-disable-next-line no-sync
+const DEFAULT_CLIENT_SECRET_HASH = cryptoService.hashPasswordSync(DEFAULT_CLIENT_SECRET);
+
 export function buildClientApplication({
   id = databaseBuffer.getNextId(),
   name = 'clientApplication',
@@ -8,16 +12,22 @@ export function buildClientApplication({
   clientSecret = 'super-secret',
   scopes = ['scope1', 'scope2'],
 } = {}) {
-  // eslint-disable-next-line no-sync
-  const hashedSecret = cryptoService.hashPasswordSync(clientSecret);
   return databaseBuffer.pushInsertable({
     tableName: 'client_applications',
     values: {
       id,
       name,
       clientId,
-      clientSecret: hashedSecret,
+      clientSecret: _getHashedSecret(clientSecret),
       scopes,
     },
   });
+}
+
+function _getHashedSecret(clientSecret) {
+  if (clientSecret === DEFAULT_CLIENT_SECRET) {
+    return DEFAULT_CLIENT_SECRET_HASH;
+  }
+  // eslint-disable-next-line no-sync
+  return cryptoService.hashPasswordSync(clientSecret);
 }