--- a/api/db/database-builder/factory/build-user.js
+++ b/api/db/database-builder/factory/build-user.js
@@ -5,8 +5,9 @@ const { isUndefined, isNil } = lodash;
 import { PIX_ADMIN } from '../../../src/authorization/domain/constants.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../src/identity-access-management/domain/constants/identity-providers.js';
 import { AuthenticationMethod, Membership } from '../../../src/shared/domain/models/index.js';
+import { DEFAULT_PASSWORD } from '../../constants.js';
 import { databaseBuffer } from '../database-buffer.js';
-import { getUserHashedPassword, USER_DEFAULT_PASSWORD } from './build-authentication-method.js';
+import { getUserHashedPassword } from './build-authentication-method.js';
 import { buildCertificationCenter } from './build-certification-center.js';
 import { buildCertificationCenterMembership } from './build-certification-center-membership.js';
 import { buildMembership } from './build-membership.js';
@@ -118,7 +119,7 @@ buildUser.withRawPassword = function buildUserWithRawPassword({
   hasSeenAssessmentInstructions = false,
   createdAt = new Date(),
   updatedAt = new Date(),
-  rawPassword = USER_DEFAULT_PASSWORD,
+  rawPassword = DEFAULT_PASSWORD,
   shouldChangePassword = false,
   emailConfirmedAt = new Date('2021-04-28T02:42:00Z'),
 } = {}) {
@@ -213,7 +214,7 @@ buildUser.withRole = function buildUserWithRole({
   createdAt = new Date(),
   updatedAt = new Date(),
   disabledAt,
-  rawPassword = USER_DEFAULT_PASSWORD,
+  rawPassword = DEFAULT_PASSWORD,
   shouldChangePassword = false,
 } = {}) {
   email = _generateEmailIfUndefined(email, id, lastName, firstName);
@@ -268,7 +269,7 @@ buildUser.withMembership = function buildUserWithMemberships({
   updatedAt = new Date(),
   organizationRole = Membership.roles.ADMIN,
   organizationId = null,
-  rawPassword = USER_DEFAULT_PASSWORD,
+  rawPassword = DEFAULT_PASSWORD,
   shouldChangePassword = false,
 } = {}) {
   email = _generateEmailIfUndefined(email, id, lastName, firstName);
@@ -369,7 +370,7 @@ buildUser.withCertificationCenterMembership = function buildUserWithCertificatio
 function _buildPixAuthenticationMethod({
   id = databaseBuffer.getNextId(),
   userId,
-  rawPassword = USER_DEFAULT_PASSWORD,
+  rawPassword = DEFAULT_PASSWORD,
   shouldChangePassword,
   createdAt,
   updatedAt,