--- a/api/db/seeds/data/team-acces/build-reset-password-users.js
+++ b/api/db/seeds/data/team-acces/build-reset-password-users.js
@@ -1,12 +1,9 @@
-import { DEFAULT_PASSWORD } from '../../../constants.js';
-
 function _buildUserWithShouldChangePassword(databaseBuilder) {
   databaseBuilder.factory.buildUser.withRawPassword({
     firstName: 'Kaa',
     lastName: 'Reboot',
     email: 'change-password@example.net',
     username: 'kaa.reboot',
-    rawPassword: DEFAULT_PASSWORD,
     shouldChangePassword: true,
     cgu: false,
   });