--- a/api/tests/tooling/domain-builder/factory/build-authentication-method.js
+++ b/api/tests/tooling/domain-builder/factory/build-authentication-method.js
@@ -7,6 +7,10 @@ import { AuthenticationMethod } from '../../../../src/identity-access-management
 import { User } from '../../../../src/identity-access-management/domain/models/User.js';
 import { cryptoService } from '../../../../src/shared/domain/services/crypto-service.js';
 
+const DEFAULT_PASSWORD = 'pix123';
+// eslint-disable-next-line no-sync
+const DEFAULT_HASHED_PASSWORD = cryptoService.hashPasswordSync(DEFAULT_PASSWORD);
+
 function _buildUser() {
   return new User({
     id: 456,
@@ -52,15 +56,13 @@ buildAuthenticationMethod.withPixAsIdentityProviderAndRawPassword = function ({
   createdAt,
   updatedAt,
 } = {}) {
-  // eslint-disable-next-line no-sync
-  const password = cryptoService.hashPasswordSync(rawPassword);
   userId = isUndefined(userId) ? _buildUser().id : userId;
 
   return new AuthenticationMethod({
     id,
     identityProvider: NON_OIDC_IDENTITY_PROVIDERS.PIX.code,
     authenticationComplement: new AuthenticationMethod.PixAuthenticationComplement({
-      password,
+      password: _getHashedPassword(rawPassword),
       shouldChangePassword,
     }),
     externalIdentifier: undefined,
@@ -141,5 +143,13 @@ buildAuthenticationMethod.withIdentityProvider = function ({
     updatedAt,
   });
 };
+
+function _getHashedPassword(password) {
+  if (password === DEFAULT_PASSWORD) {
+    return DEFAULT_HASHED_PASSWORD;
+  }
+  // eslint-disable-next-line no-sync
+  return cryptoService.hashPasswordSync(password);
+}
 
 export { buildAuthenticationMethod };