## :pancakes: Problème

Lors du chargement des seeds en local et en RA peut être un peu long parfois. Suite à une rapide investigation, il semble que le hashing de password des seeds utilisateurs prenne du temps (environ 100ms par hashing).

## :bacon: Proposition

La fonction de hashing de password est utilisé principalement à la création des seeds utilisateur. Et utilise généralement le même mot de passe pour ces utilisateurs de test. Le changement consiste à **ne pas recalculer le hash du mot de passe** par défaut des seeds à chaque fois qu'un utilisateur est créé. L'amélioration a également été appliqué sur les autres cas de seeds utilisant la fonction `cryptoService.hashPasswordSync`

**Résultats en local, Apple M3 :**
```
Avant: 11.4s en moyenne
Après: 2.3s en moyenne
Pourcentage d’amélioration: 79,82%
```

**Résultats sur les RA Scalingo :**
```
Avant: 32s en moyenne
Après: 22s en moyenne
Pourcentage d’amélioration: 31,25%
```

## 🧃 Remarques

Cette modification améliore légèrement la création des données de tests (intégration et acceptance).

## :yum: Pour tester

Faire un `time npm run db:reload` et comparer les résultats entre les branches:
- sur la branche `dev`
- sur la branche `tech-improve-database-builder-perf`